requirements.txt
# Native PyTorch for Apple Silicon
torch==2.5.1
torchvision==0.20.1

aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.45.2
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==5.0.1
attrs==23.2.0
autoflake==2.3.1
Babel==2.15.0
backoff==2.2.1
beautifulsoup4==4.12.3
black==24.10.0
bleach==6.1.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
-e git+https://github.com/PaoloTCS/ClaudeLabs.git@4639d7d07613df1ccb4824de932ee6d4ac2ddebc#egg=claudelabs
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
coverage==7.6.4
cryptography==44.0.0
cycler==0.12.1
dataclasses==0.6
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
deepsearch-glm==1.0.0
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.9
distlib==0.3.8
distro==1.9.0
dnspython==2.7.0
docling==2.14.0
docling-core==2.12.1
docling-ibm-models==3.1.0
docling-parse==3.0.0
easyocr==1.7.2
effdet==0.4.1
email_validator==2.2.0
emoji==2.14.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
executing==2.0.1
faiss-cpu==1.9.0.post1
fastapi==0.115.6
fastapi-cli==0.0.7
fastjsonschema==2.20.0
filelock==3.14.0
filetype==1.2.0
Flask==2.3.3
Flask-Compress==1.17
Flask-Cors==4.0.0
Flask-Limiter==3.10.1
flatbuffers==24.12.23
fonttools==4.53.1
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-vision==3.9.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.7
imageio==2.36.1
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.9.25
jsonlines==3.1.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.41
langchain-ollama==0.2.2
langchain-openai==0.3.7
langchain-text-splitters==0.3.4
langdetect==1.0.9
langsmith==0.2.10
layoutparser==0.3.4
lazy_loader==0.4
limits==4.0.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==2.1.5
marshmallow==3.23.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
mypy==1.13.0
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ndjson==0.3.1
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.3
nltk==3.9.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.25.2
olefile==0.47
ollama==0.4.5
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
openai==1.65.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
ordered-set==4.1.0
orjson==3.10.13
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pi_heif==0.21.0
pikepdf==9.5.1
pillow==10.4.0
pipenv==2024.0.0
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
portalocker==3.1.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.10.1
pydantic-settings==2.7.0
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pypandoc==1.14
pyparsing==3.1.2
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
pytest==8.3.3
pytest-cov==6.0.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-iso639==2024.10.22
python-json-logger==2.0.7
python-louvain==0.16
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
RapidFuzz==3.11.0
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.19.0
rsa==4.9
Rtree==1.3.0
safetensors==0.4.5
scikit-image==0.25.0
scikit-learn==1.5.1
scipy==1.14.0
semchunk==2.2.2
Send2Trash==1.8.3
sentence-transformers==3.4.1
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.12.12
tiktoken==0.9.0
timm==1.0.12
tinycss2==1.3.0
tokenizers==0.21.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typer==0.12.5
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
unstructured==0.16.12
unstructured-client==0.28.1
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.26.2
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==2.3.7
widgetsnbextension==4.0.11
wrapt==1.17.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.18.3
zstandard==0.23.0
